requirements.txt
appdirs==1.4.3
certifi==2017.4.17
chardet==3.0.3
Django==1.11.1
django-cors-headers==2.1.0
djangorestframework==3.6.4
djangorestframework-gis==0.11.2
djangorestframework-jwt==1.11.0
idna==2.5
packaging==16.8
psycopg2==2.7.1
PyJWT==1.5.2
pyparsing==2.2.0
pytz==2017.2
requests==2.17.0
six==1.10.0
urllib3==1.21.1
